Add files via upload
This project focuses on building a data ingestion pipeline using Python, similar to what was done in class with the COVID-19 cases dataset. The objective includes identifying a suitable data source with an API, describing the data, and detailing the preparation of a GET request to fetch this data, including necessary parameters and authentication type. The process involves testing the connection using Postman, writing a Python function to request the data, and then conducting data cleaning and feature engineering. Finally, the cleaned data is ingested into a PostgreSQL database, and the code is made available for review on a JupyterHub server. This project tests skills in API interaction, data handling, and database management in a Python environment.